Added reflective and dielectric materials and improved camera

+++ SV-Ray-Tracer/SV-Ray-Tracer/Utils/SV-Ray-Tracer.h
@@ -13,12 +13,12 @@ using std::sqrt;
 
 // Constants
 
-const double infinity = std::numeric_limits<double>::infinity();
+const float infinity = std::numeric_limits<float>::infinity();
 const double pi = 3.1415926535897932385;
 
 // Utility Functions
 
-inline double degrees_to_radians(double degrees) {
+inline float degrees_to_radians(float degrees) {
     return degrees * pi / 180.0;
 }
 
@@ -30,7 +30,7 @@ inline float rand_float(float min, float max) {
     return min + (max - min) * rand_float();
 }
 
-inline double clamp(double x, double min, double max) {
+inline float clamp(float x, float min, float max) {
     if (x < min) return min;
     if (x > max) return max;
     return x;
